requirement.txt
torch==1.2.0
torchvision==0.4.0
cupy==7.3.0
kornia==0.1.4
opencv-python
tqdm
scikit-image
imageio
pillow